requirements.txt
antiorm==1.2.1
anyio==4.0.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.3.0
db==0.1.1
db-sqlite3==0.0.1
exceptiongroup==1.1.3
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==13.13
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
tornado==6.3.3
typing_extensions==4.8.0
tzlocal==5.1
urllib3==2.0.6
